+++ src/core/raw_h.rs
@@ -118,8 +118,8 @@ pub fn              raw_recv       (pcb: &mut raw_pcb, raw_recv_fn recv, void *r
 #define          raw_flags(pcb) ((pcb)->flags)
 #define          raw_setflags(pcb,f)  ((pcb)->flags = (f))
 
-#define          raw_set_flags(pcb, set_flags)     do { (pcb)->flags = (u8)((pcb)->flags |  (set_flags)); } while(0)
-#define          raw_clear_flags(pcb, clr_flags)   do { (pcb)->flags = (u8)((pcb)->flags & (u8)(~(clr_flags) & 0xff)); } while(0)
+#define          raw_set_flags(pcb, set_flags)     do { (pcb)->flags = ((pcb)->flags |  (set_flags)); } while(0)
+#define          raw_clear_flags(pcb, clr_flags)   do { (pcb)->flags = ((pcb)->flags & (~(clr_flags) & 0xff)); } while(0)
 #define          raw_is_flag_set(pcb, flag)        (((pcb)->flags & (flag)) != 0)
 
 #define raw_init() /* Compatibility define, no init needed. */
