+++ src/core/tcp_h.rs
@@ -83,8 +83,8 @@ pub const TCP_FLAGS: u32 = 0x3f;U
 #define TCP_MAX_OPTION_BYTES 40
 
 #define TCPH_HDRLEN(phdr) ((lwip_ntohs((phdr)->_hdrlen_rsvd_flags) >> 12))
-#define TCPH_HDRLEN_BYTES(phdr) ((u8)(TCPH_HDRLEN(phdr) << 2))
-#define TCPH_FLAGS(phdr)  ((u8)((lwip_ntohs((phdr)->_hdrlen_rsvd_flags) & TCP_FLAGS)))
+#define TCPH_HDRLEN_BYTES(phdr) ((TCPH_HDRLEN(phdr) << 2))
+#define TCPH_FLAGS(phdr)  (((lwip_ntohs((phdr)->_hdrlen_rsvd_flags) & TCP_FLAGS)))
 
 #define TCPH_HDRLEN_SET(phdr, len) (phdr)->_hdrlen_rsvd_flags = lwip_htons(((len) << 12) | TCPH_FLAGS(phdr))
 #define TCPH_FLAGS_SET(phdr, flags) (phdr)->_hdrlen_rsvd_flags = (((phdr)->_hdrlen_rsvd_flags & PP_HTONS(~TCP_FLAGS)) | lwip_htons(flags))
