+++ src/netif/zepif.rs
@@ -199,7 +199,7 @@ zepif_linkoutput(netif: &mut netif, p: &mut pbuf)
   zep.unknown_1 = 0xff;
   zep.seq_num = lwip_htonl(state.seqno);
   state.seqno++;
-  zep.len = (u8)p.tot_len;
+  zep.len = p.tot_len;
 
   err = pbuf_take_at(q, p.payload, p.tot_len, sizeof(struct zep_hdr));
   if (err == ERR_OK) {
