+++ src/tcp_isn/tcp_isn.rs
@@ -160,10 +160,10 @@ lwip_hook_tcp_isn(const local_ip: &mut ip_addr_t, local_port: u16,
   }   
 
 
-  input[32] = (u8)(local_port >> 8);
-  input[33] = (u8)(local_port & 0xff);
-  input[34] = (u8)(remote_port >> 8);
-  input[35] = (u8)(remote_port & 0xff);
+  input[32] = (local_port >> 8);
+  input[33] = (local_port & 0xff);
+  input[34] = (remote_port >> 8);
+  input[35] = (remote_port & 0xff);
 
   /* The secret and padding are already filled in. */
